+++ packages/complex-rubric/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/complex-rubric@2.3.5...@pie-element/complex-rubric@2.4.0) (2024-03-04)
+
+
+### Features
+
+* update pie-lib/pie-toolbox version PD-1870 and PD-3501 ([8df1d8d](https://github.com/pie-framework/pie-elements/commit/8df1d8d281a0d6023d5a9c0e6fdd29fea71b6ac1))
+
+
+
+
+
+## [2.3.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/complex-rubric@2.3.4...@pie-element/complex-rubric@2.3.5) (2024-02-29)
+
+**Note:** Version bump only for package @pie-element/complex-rubric
+
+
+
+
+
 ## [2.3.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/complex-rubric@2.3.3...@pie-element/complex-rubric@2.3.4) (2024-02-22)
 
 **Note:** Version bump only for package @pie-element/complex-rubric