+++ packages/complex-rubric/configure/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.18](https://github.com/pie-framework/pie-elements/compare/@pie-element/complex-rubric-configure@2.1.17...@pie-element/complex-rubric-configure@2.1.18) (2024-03-04)
+
+**Note:** Version bump only for package @pie-element/complex-rubric-configure
+
+
+
+
+
+## [2.1.17](https://github.com/pie-framework/pie-elements/compare/@pie-element/complex-rubric-configure@2.1.16...@pie-element/complex-rubric-configure@2.1.17) (2024-02-29)
+
+**Note:** Version bump only for package @pie-element/complex-rubric-configure
+
+
+
+
+
 ## [2.1.16](https://github.com/pie-framework/pie-elements/compare/@pie-element/complex-rubric-configure@2.1.15...@pie-element/complex-rubric-configure@2.1.16) (2024-02-22)
 
 **Note:** Version bump only for package @pie-element/complex-rubric-configure